# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EmbodiedCloud API",
    "description": "Browser-first Isaac Lab workspace control plane",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/courses": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Courses",
        "operationId": "list_courses_api_courses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CourseOut"
                  },
                  "type": "array",
                  "title": "Response List Courses Api Courses Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Create Course",
        "operationId": "create_course_api_courses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Course",
        "operationId": "get_course_api_courses__course_id__get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/join": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Join Course",
        "operationId": "join_course_api_courses__course_id__join_post",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseMemberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/join-by-slug": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Join Course By Slug",
        "description": "\u6309 slug\uff08\u9080\u8bf7\u7801\uff09\u52a0\u5165\u8bfe\u7a0b\uff1a\u8001\u5e08\u5206\u4eab slug\uff0c\u5b66\u751f\u8f93\u5165\u5373\u53ef\u52a0\u5165\u3002",
        "operationId": "join_course_by_slug_api_courses_join_by_slug_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinBySlugIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseMemberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/members": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Add Member",
        "operationId": "add_member_api_courses__course_id__members_post",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberAdd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseMemberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Members",
        "operationId": "list_members_api_courses__course_id__members_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseMemberOut"
                  },
                  "title": "Response List Members Api Courses  Course Id  Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/completions": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get Completions",
        "operationId": "get_completions_api_courses__course_id__completions_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/labs": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Create Lab",
        "operationId": "create_lab_api_courses__course_id__labs_post",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Labs",
        "description": "labs \u5217\u8868\u5bf9 member \u53ef\u89c1\uff08\u5b66\u751f\u8981\u770b\u5230\u5b9e\u9a8c\u624d\u80fd launch\uff09\uff1b\u521b\u5efa\u4ecd\u9650 teacher\u3002",
        "operationId": "list_labs_api_courses__course_id__labs_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LabOut"
                  },
                  "title": "Response List Labs Api Courses  Course Id  Labs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses/{course_id}/my-progress": {
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "Get My Progress",
        "description": "\u5f53\u524d\u7528\u6237\u5728\u8bfe\u7a0b\u4e2d\u7684\u4f5c\u4e1a\u8fdb\u5ea6\uff08member \u53ef\u89c1\uff1b\u6559\u5e08\u53e6\u7528 /completions \u770b\u5168\u73ed\uff09\u3002",
        "operationId": "get_my_progress_api_courses__course_id__my_progress_get",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/labs/{lab_id}/assignments": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Create Assignment",
        "operationId": "create_assignment_api_labs__lab_id__assignments_post",
        "parameters": [
          {
            "name": "lab_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lab Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "courses"
        ],
        "summary": "List Assignments",
        "description": "assignments \u5217\u8868\u5bf9 member \u53ef\u89c1\uff08\u5b66\u751f\u8981\u770b\u5230\u4f5c\u4e1a\u624d\u80fd\u63d0\u4ea4\uff09\uff1b\u521b\u5efa\u4ecd\u9650 teacher\u3002",
        "operationId": "list_assignments_api_labs__lab_id__assignments_get",
        "parameters": [
          {
            "name": "lab_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lab Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentOut"
                  },
                  "title": "Response List Assignments Api Labs  Lab Id  Assignments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/labs/{lab_id}/launch": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Launch Lab",
        "operationId": "launch_lab_api_labs__lab_id__launch_post",
        "parameters": [
          {
            "name": "lab_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lab Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/assignments/{assignment_id}/submit": {
      "post": {
        "tags": [
          "courses"
        ],
        "summary": "Submit Assignment",
        "operationId": "submit_assignment_api_assignments__assignment_id__submit_post",
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assignment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignmentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "AssignmentCreate"
      },
      "AssignmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "lab_id": {
            "type": "string",
            "title": "Lab Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "lab_id",
          "name",
          "description",
          "due_at",
          "created_at"
        ],
        "title": "AssignmentOut"
      },
...
      "CourseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "slug": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "title": "CourseCreate"
      },
      "CourseMemberOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "user_id",
          "role"
        ],
        "title": "CourseMemberOut"
      },
      "CourseOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "owner_id",
          "name",
          "description",
          "slug",
          "created_at"
        ],
        "title": "CourseOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JoinBySlugIn": {
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Slug"
          }
        },
        "type": "object",
        "required": [
          "slug"
        ],
        "title": "JoinBySlugIn"
      },
      "LabCreate": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "quota_seconds": {
            "type": "integer",
            "title": "Quota Seconds",
            "default": 3600
          }
        },
        "type": "object",
        "required": [
          "template_id",
          "name"
        ],
        "title": "LabCreate"
      },
      "LabOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "title": "Course Id"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "quota_seconds": {
            "type": "integer",
            "title": "Quota Seconds"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "template_id",
          "name",
          "description",
          "quota_seconds",
          "created_at"
        ],
        "title": "LabOut"
      },
...
      "MemberAdd": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "student"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "MemberAdd"
      },
...
      "SubmissionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "assignment_id": {
            "type": "string",
            "title": "Assignment Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "assignment_id",
          "user_id",
          "workspace_id",
          "status",
          "created_at",
          "completed_at"
        ],
        "title": "SubmissionOut"
      },
      "SubmitIn": {
        "properties": {
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "type": "object",
        "title": "SubmitIn"
      },
...
      "WorkspaceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "gpu_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Index"
          },
          "gpu_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Name"
          },
          "gpu_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Id"
          },
          "ide_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ide Url"
          },
          "stream_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Hint"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "stopped_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopped At"
          },
          "accumulated_seconds": {
            "type": "integer",
            "title": "Accumulated Seconds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "template_id",
          "user_id",
          "organization_id",
          "status",
          "provider",
          "gpu_index",
          "gpu_name",
          "gpu_id",
          "ide_url",
          "stream_hint",
          "error_message",
          "created_at",
          "started_at",
          "stopped_at",
          "accumulated_seconds"
        ],
        "title": "WorkspaceOut"
      }
    }
  }
}